backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
nuscenes-devkit==1.1.11
numpy==1.24.3
Pillow==10.2.0
